WLTS-28: Refinement for version id and blank translations with manage-traslation role           [WLTS-28] (https://tools.hmcts.net/jira/browse/WLTS-28).

+++ src/integrationTest/java/uk/gov/hmcts/reform/translate/controllers/DictionaryControllerIT.java
@@ -376,7 +376,7 @@ public class DictionaryControllerIT extends BaseTest {
         return new Dictionary(expectedMapKeysAndValues);
     }
 
-    private Integer assertDictionaryEntityWithTranslationPhrases(String englishPhrase) {
+    private Long assertDictionaryEntityWithTranslationPhrases(String englishPhrase) {
         val dictionaryEntity = dictionaryRepository.findByEnglishPhrase(englishPhrase);
         assertTrue(dictionaryEntity.isPresent());
         assertNotNull(dictionaryEntity.get().getTranslationUpload());
